cdfs: renamed driver to cdfs2 and changed priority to avoid conflicts
The system already comes with this same cdfs driver, so we need
to make these tweaks so that we can override the system driver.
Note that it is not possible to overwrite the original system
driver because of system file protection (and it's not possible to
unload the driver without rebooting).

Signed-off-by: Vitaly Chipounov <vitaly@cyberhaven.io>